% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bayz}
\alias{coef.bayz}
\title{Extract posterior means and SDs for all coefficients from a bayz model
output}
\usage{
\method{coef}{bayz}(bayz_output, ...)
}
\arguments{
\item{...}{Additional parameters.}

\item{object}{A bayz model output}
}
\value{
A data.frame with estimates for all model coefficients
}
\description{
Rbayz::coef() extracts estimates (posterior mean and SD) for all
coefficients from the bayz output. Coefficients is here defined as all fixed
and random effects, i.e. "regression coefficients" in a broad meaning.
The coef() function is a wrapper around estim() and 'unlists' the output to
return a single data frame.
Other functions to extract estimates from the bayz output are fixef(),
ranef(), vhest(), and the generic estim().
}
