% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{estim}
\alias{estim}
\title{Extract parameter estimates from a bayz model output}
\usage{
estim(object, param = NULL, splitLabels = FALSE, unlist = FALSE, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{param}{A string with the name of the parameter for which to
retrieve estimates, or a vector of strings with the
names of multiple parameters. This argument can be
omitted to retrieve all parameter estimates.}

\item{splitLabels}{Whether labels in interactions should be split in
multiple columns (default FALSE). Splitting labels is
only possible when retrieving estimates for a single
parameter.}

\item{unlist}{Whether to return multiple sets of parameter estimates
in a single data frame
with an additional column Parameter to indicate from
which set of parameters each estimate comes. If FALSE
(default) multiple parameters are returned in a list of
data frames.}

\item{...}{Additional parameters.}
}
\value{
Dependent on usage, a single data frame with parameter estimates,
        or a list of data frames with estimates for multiple parameters.
}
\description{
Rbayz::estim() extracts one, several, or all parameter estimates from a
bayz output, reporting posterior means and standard deviations. Parameter
estimates also can be accessed directly in the $Estimates list in the
output, but the estim() function can split labels of interaction terms,
combine several parameters in one output, and can unlist the output, merging
estimates from several parameters in a single data frame.
Main uses are:
- estim(output) returns all parameter estimates as a list of data
  frames. The estimates for one set of parameters could then be accessed as
  estim(output)$`parameter-name`, for example estim(output)$"year.location".
- estim(output, param=name) with param set to a string returns a single data
  frame with
  estimates for the requested parameter. Example: estim(output, 
  param="year.location"), or shortly estim(output, "year.location").
- estim(output, param=c(name1, name2, ...)) with param a vector of strings
  returns a list of data frames with estimates for the requested parameters.
For the first and third usage, adding unlist=TRUE will reformat the default
list output into a single data frame.
For the second usage, adding splitlabels=TRUE will split labels of
interaction terms in multiple columns, which can be easier
to read and match with the original data. The second and third usage has
better error handling because estim() checks exsistence of the requested
parameter(s), and returns NULL if not available; with the first usage, and
when directly accessing $Estimates in the output, one risks attempting to
access non-existing parameters.
Parameter names are based on the variable names in the model-term, but can
have small modifications: colons in interactions
are replaced by dots; and multiple use of the same variable can have
appended an extra digit. Random effect terms will create two slots in the
parameter list, one for the random effects themselves
and one for the variance (or other) hyper-parameters. If in doubt about
parameter names, check the $Parameters data frame in the output. 
Apart from estim(), Rbayz also provides fixef() and ranef() to retrieve
fixed and random effects, respectively, coef() to retrieve the combined
fixed and random effects, and
vhest() to retrieve estimates for variance and hyper-parameters.
}
