% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef.bayz}
\alias{fixef.bayz}
\title{Extract posterior means and SDs for "fixed" effects from a bayz model output}
\usage{
\method{fixef}{bayz}(object, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{...}{Additional parameters passed onto the Model function.}
}
\value{
a data frame with estimates for all fixed effects
}
\description{
The fixef() function extracts estimates (posterior mean and SD) for all
fixed effects from the bayz output. Fixed effects are here defined as the
coefficients for the intercept and fx() and rg() model terms.
The fixef() function is a wrapper around estim() and 'unlists' the output to
return a single data frame.
To extract estimates for one, or a subset, of the fixed effects, use the
more generic estim(). Use ranef() to extract estimates for all random
effects.
}
