% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.conv.R, R/print.summary.R
\name{print.summarybayz}
\alias{print.summarybayz}
\title{Print function for 'summarybayz' class}
\usage{
\method{print}{summarybayz}(object, ...)

\method{print}{summarybayz}(object, ...)
}
\arguments{
\item{...}{additional parameters}

\item{x}{bayz object}
}
\value{
Nothing, only prints

Nothing, only prints
}
\description{
Prints the bayz class

Prints the bayz class
}
