% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prval.R
\name{prval}
\alias{prval}
\title{Compute pr-values (p-values from random effects) from a bayz model output}
\usage{
prval(object, param = NULL, splitlabels = FALSE, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{param}{Parameter name for which to extract r-values (if in
doubt about naming, check $Parameters from the output)}

\item{splitlabels}{If the estimates come from an interaction, whether
labels should be split to allow easy matching to levels
in the single variables (default FALSE).}

\item{...}{Additional parameters.}
}
\value{
a data frame with parameter estimates (posterior mean and SD), the
                     backtransformed z-statistic and its pr (p) -value.
}
\description{
Rbayz::prval() evaluates significances from random effects based on
procedures from Gualdron-Duarte et al. (2013)
and Bernal-Rubio et al. (2016), extended to allow use in a full Bayesian
framework where the variance hyper-parameters are also learned from the data.
The G-D & B-R procedure applies to fitting high-dimensional predictor sets
as random effects, as done in bayz with rr(),
and evaluates the significance of individual predictors
as if they were fitted one by one in turn as fixed in a linear mixed model
with all other effects remaining random.
The reported pr-value is based on a z-test.
Note that the model-term for which pr-values are to be computed should have
set a 'save' or 'prval' option.
}
