% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.bayz}
\alias{ranef.bayz}
\title{Extract posterior means and SDs for "random" effects from a bayz model output}
\usage{
\method{ranef}{bayz}(object, ...)
}
\arguments{
\item{object}{A bayz model output.}
}
\value{
A data frame with estimates for all random effects.
}
\description{
The ranef() function extracts estimates (posterior mean and SD) for all
random effects from the bayz output. Random effects are here defined as the
coefficients for the rn() and rr() model terms.
The ranef() function is a wrapper around estim() and 'unlists' the output to
return a single data frame.
To extract estimates for one, or a subset, of the random effects, use the
more generic estim(). Use fixef() to extract estimates for all fixed
effects.
}
