% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R, R/summary.R
\name{summary.bayz}
\alias{summary.bayz}
\title{Convergence diagnostics on traced parameters from bayz model fit}
\usage{
\method{summary}{bayz}(object, HPDprob = 0.95, burnin = NULL, ...)

\method{summary}{bayz}(object, HPDprob = 0.95, burnin = NULL, ...)
}
\arguments{
\item{object}{bayz output object}

\item{HPDprob}{probability for the Highest Posterior Density intervals
(default 0.95)}

\item{burnin}{Discard additional samples with cycle number less or equal to this (updated) burnin setting.
Setting burnin here lower than the original chain burnin has no effect.}

\item{...}{additional parameters}
}
\value{
bayzconv object; if bayz finished without errors, a list with a convergence table and convergence status,
        otherwise a list with error messages.

summarybayz object
}
\description{
Rbayz::conv() computes convergence diagnostics on the 'traced' (samples-saved) parameters from a bayz model fit:
effective sample size, Geweke Z-value, Monte Carlo Standard Error (MCSE) and Monte Carlo Coefficient of
Variation (MCCV%) using functions from the coda package. Note that the MCCV% can remain high when the parameter's
posterior mean is close to zero and the posterior mean is therefore also reported to help interpret the MCCV%.
Parameter 'burnin' can be used to increase the burnin setting from the orginal chain and discard additional samples
from the start of the chain.

Summary() presents an overview of the model fitted, any errors and warnings,
and posterior summary statistics, including HPD intervals, for the main model
parameters in a bayz output. Summary() collects the 'traced' parameters,
for which samples are saved in the output, allowing to compute HPD intervals
(as well as trace plots, density plots and convergence diagnostics as done by
plot() and conv()). The 'traced' parameters are by default intercept,
fixed effects and regressions with up to 4 levels, and variances
and hyper-paramters. Tracing can be user-toggled
with a 'trace' option on model terms. Computation of HPD intervals accounts
for parameter boundaries, allowing variance intervals starting at zero.
The conv() function computes convergence diagnostics on the same set of
traced parameters.
}
